Convert Narrowphase to TS
